requirements.txt
Django==3.1.7
django-rest-framework==0.1.0
djangorestframework==3.12.2
drf-access-policy==0.8.7

